requirements.txt
requests==2.31.0
python-telegram-bot==20.7
watchdog==3.0.0
redis==5.0.1
flask==3.0.0
python-dotenv==1.0.0
Pillow==10.1.0
